api_db: Add working skeleton code for api_db, add dockerfile
Added skeleton code for api_db, which consists of FastAPI routes and app.py with an entry point
in the main function. This app currently consists of some health check endpoints and  an endpoint to check environment configuration
Also added a dockerfile which can be used to build a working docker image of the api_db app.

Bug: T284399
Change-Id: I96f622ebe8e3f876145a1d4b5b222e333b79724c